Cocoa

Objective-C like Type mappings

type NSUInteger = CULong

just to make things read a bit clearer

Treat NSUInteger like CULong.. this is 64-bit dependent.

type NSInteger = CInt

type NSBool = CInt

newtype Class

Constructors

Class (Ptr ())

newtype Selector

Constructors

Selector (Ptr ())

newtype Method

Constructors

Method (Ptr ())

newtype Id

Constructors

Id (Ptr ())

data NSURL a

Constructors

NSURL Id

data NSString a

Constructors

NSString Id

nsYES :: NSBool

YES in obj-c. this is a Int with the value 1

nsNO :: NSBool

NO in obj-c. this is a Int with the value 0

FFI imports

class_getClassMethod :: Class -> Selector -> IO Method

ffi import for class_getClassMethod

objc_getClass :: CString -> IO Id

ffi import for objc_getClass

sel_registerName :: CString -> IO Selector

ffi import for sel_registerName

objc_msgSend :: Id -> Selector -> IO Id

These are defined via the objch.c, as ffi + varargs is a world of pain |ffi import for objc_msgSend taking no arguments

objc_msgSend1 :: Id -> Selector -> Id -> IO Id

ffi import for objc_msgSend taking 1 argument

objc_msgSend2 :: Id -> Selector -> Id -> Id -> IO Id

ffi import for objc_msgSend taking 2 arguments

objc_msgSend3 :: Id -> Selector -> Id -> Id -> Id -> IO Id

ffi import for objc_msgSend taking 3 arguments

objc_msgSend4 :: Id -> Selector -> Id -> Id -> Id -> Id -> IO Id

ffi import for objc_msgSend taking 4 arguments

objc_msgSend5 :: Id -> Selector -> Id -> Id -> Id -> Id -> Id -> IO Id

ffi import for objc_msgSend taking 5 arguments

objc_msgSend6 :: Id -> Selector -> Id -> Id -> Id -> Id -> Id -> Id -> IO Id

ffi import for objc_msgSend taking 6 arguments

objc_msgSend7 :: Id -> Selector -> Id -> Id -> Id -> Id -> Id -> Id -> Id -> IO Id

ffi import for objc_msgSend taking 7 arguments

objc_msgSend8 :: Id -> Selector -> Id -> Id -> Id -> Id -> Id -> Id -> Id -> Id -> IO Id

ffi import for objc_msgSend taking 8 arguments

nilPtr :: Id

nil in objective-c

id2nsinteger :: Id -> IO NSInteger

hacks because objc_msgSend can return primitives, instead of the id struct..

nsInteger2id :: NSInteger -> IO Id

id2nsuinteger :: Id -> IO NSUInteger

nsuinteger2id :: NSUInteger -> IO Id

id2bool :: Id -> IO NSBool

bool2id :: NSBool -> IO Id

id2cstr :: Id -> IO CString

cstr2id :: CString -> IO Id

Type classes

class NSObject a where

Used to create light wrappers around objects that are based Id this could be used for any NSObject object. its only used for NSString at the moment so overloaded strings can auto convert to NSString (and Data.Eq to compare)

Methods

idVal :: a -> Id

The underlying Id object

Functions

msgSend :: Id -> Selector -> [Id] -> IO Id

Sends Selector to Id. a message to an object.

getClass :: String -> IO Id

Returns a class Id for the given String.

getClass' :: String -> Id

Returns a class Id for the given String. this function uses unsafePerformIO

registerName :: String -> IO Selector

registerName' :: String -> Selector

getStringId :: NSString a -> Id

Retrieve the underlying Id type from an NSObject

getURLId :: NSURL a -> Id

Retrieve the underlying Id type from an NSObject

newNSString :: String -> IO (NSString Id)

Create a new NSString

newNSString' :: String -> NSString Id

Create a new NSString this function uses unsafePerformIO

nsboolToBool :: NSBool -> Bool

Convert an NSBool to Bool

nsstringToString :: NSString Id -> IO String

Convert an NSString to an IO String

isEqualToString :: NSString Id -> NSString Id -> IO Bool

Checks equality of two NSString, uses isEqualToString: internally.

isEqualToString' :: NSString Id -> NSString Id -> Bool

Checks equality of two NSString, uses isEqualToString: internally. this function uses unsafePerformIO

newAutoreleasePool :: IO Id

Allocates a new autorelease pool, newAutoreleasePool is the equivalent of [[NSAutoreleasePool alloc] init];

delAutoreleasePool :: Id -> IO ()

Deallocates an autorelease pool, delAutoreleasePool is the equivalent of [pool release];

startRunloop :: IO ()

This is required to make obj-c calls work enables the default runloop, this call locks the thread its run in, so its advised to forkIO it. This is the equivalent to [[NSRunloop mainRunLoop] run];

nsObjectDescribe :: Id -> IO String

Returns a string representation of the current obj-c object. This is the equivalent of [x description];

NSArray Functions

arrayCount :: Id -> IO NSUInteger

Items in array equiv to: [array count];

arrayObjectAtIndex :: Id -> NSUInteger -> IO Id

Retrieve object at specified index equiv to: [array objectAtIndex:index];

arrayToList :: Id -> IO [Id]

Convert NSArray to list

NSURL Functions

fileURLWithPath :: NSString Id -> IO (NSURL Id)

Specify path to file returning an NSURL. equiv to: [NSURL fileURLWithPath:@""];

fileURLPath :: NSURL Id -> IO String

Retrieve path specified by the NSURL as a string. equiv to: [url path];

NSData Functions

dataWithContentsOfURL :: Id -> IO Id

Return NSData from path specified by an NSURL. equiv to: [NSData dataWithContentsOfURL:url];

Infix operators

($<-) :: Id -> Selector -> IO Id

Infix for msgSend with an empty list as args.

($<<-) :: Id -> Selector -> [Id] -> IO Id

Infix for msgSend

($<~) :: IO Id -> IO Selector -> IO Id

Infix for msgSend taking values wrapped with IO with an empty list as args

($<<~) :: IO Id -> IO Selector -> [IO Id] -> IO Id

Infix for msgSend taking values wrapped with IO