Cocoa.NSObject

Description

NSObject is the root class of most Objective-C class hierarchies. Through NSObject, objects inherit a basic interface to the runtime system and the ability to behave as Objective-C objects.

Documentation

description :: IsNSObject a => a -> IO String

Returns a string representation of the current obj-c object. This is the equivalent of [x description];