Cocoa.NSString

Description

The NSString class declares the programmatic interface for an object that manages immutable strings. An immutable string is a text string that is defined when it is created and subsequently cannot be changed. NSString is implemented to represent an array of Unicode characters, in other words, a text string.

Types

data TyNSString a

IsNSObject compatable wrapper for NSStrings

Constructors

TyNSString Id

Functions

newNSString :: String -> IO (TyNSString Id)

Create a new NSString

newNSString' :: String -> TyNSString Id

Create a new NSString this function uses unsafePerformIO

isEqualToString :: TyNSString Id -> TyNSString Id -> IO Bool

Checks equality of two NSString, uses isEqualToString: internally.

isEqualToString' :: TyNSString Id -> TyNSString Id -> Bool

Checks equality of two NSString, uses isEqualToString: internally. this function uses unsafePerformIO

getStringId :: TyNSString a -> Id

Retrieve the underlying Id type from an NSObject

toNSString :: IO Id -> IO (TyNSString Id)

Wrap IO Id object in TyNSString

nsstringToString :: TyNSString Id -> IO String

Convert an NSString to an IO String