Cocoa.PdfKit

Types

type PDFDocument = Id

type PDFPage = Id

type PDFAnnotation = Id

type PDFAnnotationButtonWidget = PDFAnnotation

type PDFAnnotationCircle = PDFAnnotation

type PDFAnnotationFreeText = PDFAnnotation

type PDFAnnotationInk = PDFAnnotation

type PDFAnnotationLine = PDFAnnotation

type PDFAnnotationLink = PDFAnnotation

type PDFAnnotationMarkup = PDFAnnotation

type PDFAnnotationPopup = PDFAnnotation

type PDFAnnotationSquare = PDFAnnotation

type PDFAnnotationStamp = PDFAnnotation

type PDFAnnotationText = PDFAnnotation

type PDFAnnotationTextWidget = PDFAnnotation

PdfKit Functions

Initializing Documents

initWithData :: Id -> IO PDFDocument

initWithURL :: TyNSURL Id -> IO PDFDocument

Accessing Document Information

documentURL :: PDFDocument -> IO (TyNSURL Id)

Returns the URL for the document.

majorVersion :: PDFDocument -> IO NSInteger

Returns the major version of the document.

minorVersion :: PDFDocument -> IO NSInteger

Returns the minor version of the document.

string :: PDFDocument -> IO (TyNSString Id)

Returns a string representing the textual content for the entire document.

Managing Document Security

isEncrypted :: PDFDocument -> IO NSBool

Returns a Boolean value specifying whether the document is encrypted.

isLocked :: PDFDocument -> IO NSBool

Returns a Boolean value indicating whether the document is locked.

Writing Out the PDF Data

writeToFile :: PDFDocument -> TyNSString Id -> IO NSBool

Writes the document to a file at the specified path.

Working with Pages

pageCount :: PDFDocument -> IO NSUInteger

Returns the number of pages in the document.

pageAtIndex :: PDFDocument -> NSUInteger -> IO PDFPage

Returns the page at the specified index number.

indexForPage :: PDFDocument -> PDFPage -> IO NSUInteger

Gets the index number for the specified page.

insertPageAtIndex :: PDFDocument -> PDFPage -> NSUInteger -> IO ()

Inserts a page at the specified index point.

removePageAtIndex :: PDFDocument -> NSUInteger -> IO ()

Managing Find Operations

Working with Selections

Setting the Delegate

Searching Documents

Unlocking Documents

Determining the Page Class

PdfPage Functions

Initializing a Page

Getting Information About a Page

Working with Annotations

annotations :: PDFPage -> IO [PDFAnnotation]

Returns an array containing the page’s annotations.

Rendering Pages

Working with Textual Content

Working with Selections

Miscellaneous

PdfAnnotation Functions

Initializing an Annotation

Accessing Information About an Annotation

annotationType :: PDFAnnotation -> IO (TyNSString Id)

Returns the type of the annotation. Types include Line, Link, Text, and so on

Managing Annotation Display Characteristics

Managing Annotation Drawing and Output

Misc

annotationIsLink :: PDFAnnotation -> IO Bool

annotationLinks :: PDFPage -> IO [PDFAnnotationLink]

annotationLinkURL :: PDFAnnotationLink -> IO String