Swizzle

Description

Replace objective-c methods with haskell functions at runtime.

 {-# NOINLINE swizzleDescription #-}
 swizzleDescription :: SwizzleVoid
 swizzleDescription _ _ = do
    putStrLn "No soup for you"

... 
 description str >>= putStrLn
 str <- toNSString $ ("NSDate" $<- "alloc") >>= str -> str $<- "init"
 swizzleHSMethod (idVal str) "description" swizzleDescription
 description str >>= putStrLn

Outputs:

2013-09-16 14:37:28 +0000

No soup for you

Types

type SwizzleVoid = Id -> Selector -> IO ()

Represents an objective-c method

FFI Imports

class_addMethod' :: Class -> Selector -> FunPtr SwizzleVoid -> CString -> Bool

Adds a new method to a Class with a given name and implementation.

mkSwizzleVoid :: SwizzleVoid -> IO (FunPtr SwizzleVoid)

Wrapper that converts a SwizzleVoid to a FunPtr SwizzleVoid

Functions

swizzle :: Class -> Selector -> Selector -> IO ()

Replaces original Selector on Class with a new Selector

swizzleMethod :: Id -> Selector -> Selector -> IO ()

Replaces original Selector on instance of an object with the second Selector

swizzleHSMethod :: Id -> Selector -> SwizzleVoid -> IO ()

Replaces original Selector on instance of an object with haskell SwizzleVoid